refact: move Stack measuring and placing to StackMeasurePolicy
